pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "serial-mcp-server"
dynamic = ["version"]
description = "Serial port Model Context Protocol server for AI agents and developer tooling"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
dependencies = [
    "pyserial>=3.5,<4",
    "mcp>=1.0,<2",
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]
dev = [
    "pre-commit>=4.0",
    "ruff>=0.9",
]

[project.scripts]
serial_mcp = "serial_mcp_server.server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["serial_mcp_server"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.ruff]
target-version = "py311"
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = [
    "E501",    # line length handled by formatter
    "SIM105",  # contextlib.suppress is less readable for async cleanup
    "RUF001",  # en-dash in strings is intentional
    "RUF002",  # en-dash in docstrings is intentional
    "RUF003",  # en-dash in comments is intentional
    "RUF006",  # fire-and-forget create_task is intentional in signal handlers
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["RUF059"]  # unused unpacked vars in tests are fine

[tool.ruff.lint.isort]
known-first-party = ["serial_mcp_server"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "serial_mcp_server/_version.py"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B110"]  # try-except-pass is intentional for best-effort serial cleanup
